Создан абстрактный класс Component, общие свойства классов Point/PointEdit вынесены в него
